Refactoring and code overhaul
I decided to redesign my code, clean it up a bit and start doing the
conveyor logic from scratch, it was too janky. It is still janky but at
least items don't go into each other. I'll still need to refactor this
later on though

Amended: It is no longer janky, but because I didn't edit too much I
decided to edit this commit instead of creating a new commit with little
modifications